Rename cp to control_plane
